Moved giveClasses method into PlayerClassManager
