+++ src/main/java/lar/minecraft/hg/managers/ServerManager.java
@@ -40,43 +40,6 @@ public class ServerManager {
 		return livingPlayers;
 	}
 		
-	public static void giveClasses() {
-		for(Player player : getLivingPlayers()) {
-			PlayerInventory playerInventory = player.getInventory();
-			playerInventory.clear();
-			playerInventory.addItem(new ItemStack(Material.COMPASS));
-			PlayerExtra playerExtra = SpigotPlugin.playerExtras.getOrDefault(player.getUniqueId(), null);
-			if (playerExtra != null && PlayerClassManager.AVAILABLE_PLAYER_CLASSES.containsKey(playerExtra.getPlayerClass().getName())) {
-				playerExtra.getPlayerClass().getAction().perform(player);
-		        /*
-				switch (chosenClass) {
-					case "bowman":
-						playerInventory.addItem(new ItemStack(Material.BOW), new ItemStack(Material.ARROW, 16));
-						break;
-					case "armored":
-						playerInventory.setHelmet(new ItemStack(Material.IRON_HELMET));
-						playerInventory.setChestplate(new ItemStack(Material.IRON_CHESTPLATE));
-						playerInventory.setLeggings(new ItemStack(Material.IRON_LEGGINGS));
-						playerInventory.setBoots(new ItemStack(Material.IRON_BOOTS));
-						break;
-					case "doglover":
-						playerInventory.addItem(new ItemStack(Material.BONE, 8));
-						Wolf dog = SpigotPlugin.server.getWorld("world").spawn(player.getLocation(), Wolf.class);
-						dog.setOwner(player);
-						dog.setAdult();
-						dog.setSitting(false);
-						break;
-					case "lavaman":
-						playerInventory.addItem(new ItemStack(Material.LAVA_BUCKET), new ItemStack(Material.LAVA_BUCKET), new ItemStack(Material.LAVA_BUCKET));
-						break;
-					default:
-						break;
-				}*/
-			}
-			player.updateInventory();
-		}
-	}
-	
 	public static void restartServer() {
 		// Save all worlds before restarting
 		Bukkit.getScheduler().runTask(SpigotPlugin.getPlugin(SpigotPlugin.class), () -> {
